+++ src/main/kotlin/no/nav/syfo/common/types/ident/Navident.kt
@@ -1,6 +1,6 @@
 package no.nav.syfo.common.types.ident
 
 @JvmInline
-public value class NavIdent(
+public value class Navident(
     public val value: String,
 )