+++ src/testFixtures/kotlin/no/nav/syfo/common/mock/token/azuread/MockAzureAdRequestHandler.kt
@@ -10,7 +10,7 @@ import no.nav.syfo.common.mock.respond
  * so that the token can be read out again in mock request handlers for other services that are called with an obo token,
  * like [mockTilgangskontrollRequestHandler][no.nav.syfo.common.mock.tilgangskontroll.mockTilgangskontrollRequestHandler].
  * These mock handlers can use the token to read the NAV-ident claim and simulate different behavior like different
- * tilgang for different mock users / navIdents.
+ * tilgang for different mock users / navidents.
  *
  * This will respond with the incoming "assertion token" as accessToken for getOnBehalfOfToken calls,
  * and with "token" as the accessToken for getSystemToken calls.